+++ packages/jshark-lucid/bench/Main.hs
@@ -9,6 +9,7 @@ import qualified Data.Text as T
 import JShark
 import JShark.Api
 import JShark.Api.Rec (Rec (..))
+import JShark.Internal (effectfulAST, optimizedEffectSize)
 import JShark.Lucid
 import Lucid (button_, class_, div_, label_, li_, type_)
 import System.CPUTime