+++ packages/jshark/bench/Main.hs
@@ -8,6 +8,7 @@ import Bench.Stages (codepathStages, codepathStagesPure)
 import JShark
 import JShark.Api
 import JShark.Api.Rec (Rec (..), (<:))
+import JShark.Internal (optimizedEffectSize)
 import Test.Tasty.Bench
 
 -- A long chain of binds: